LPS-172038 portal-search-web: Pull up testEmptySearchResults() into BaseTestCase
https://issues.liferay.com/browse/LPS-172038
